Ädded : Mac OS Rive Libraries

+++ Source/RiveCore/Private/Assets/URFile.cpp
@@ -58,13 +58,13 @@ void UE::Rive::Assets::FURFile::PrintStats()
 
     RiveFileLoadArgs.Add(TEXT("Minor"), FText::AsNumber(NativeFilePtr->minorVersion));
 
-    RiveFileLoadArgs.Add(TEXT("NumArtboards"), FText::AsNumber(NativeFilePtr->artboardCount()));
+    RiveFileLoadArgs.Add(TEXT("NumArtboards"), FText::AsNumber((int32)NativeFilePtr->artboardCount()));
 
-    RiveFileLoadArgs.Add(TEXT("NumAssets"), FText::AsNumber(NativeFilePtr->assets().size()));
+    RiveFileLoadArgs.Add(TEXT("NumAssets"), FText::AsNumber((int32)NativeFilePtr->assets().size()));
 
     if (const rive::Artboard* NativeArtboard = NativeFilePtr->artboard())
     {
-        RiveFileLoadArgs.Add(TEXT("NumAnimations"), FText::AsNumber(NativeArtboard->animationCount()));
+        RiveFileLoadArgs.Add(TEXT("NumAnimations"), FText::AsNumber((int32)NativeArtboard->animationCount()));
     }
     else
     {
